+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+-   Hide the module checkboxes if the module is not selected
+-   ESLint update - allow unused vars that start with an underscore
 
 ## [2.0.5] - 2023-02-20